Revert "New routes"
