--- components/TestCard.vue
@@ -1,15 +1,0 @@
-<template>
-  <div></div>
-</template>
-
-<script>
-import { mapGetters } from 'vuex'
-
-export default {
-  computed: {
-    ...mapGetters(['currentTest'])
-  }
-}
-</script>
-
-<style lang="scss" scoped></style>
