+++ components/delete?.vue
@@ -125,6 +125,7 @@ export default {
         incorrect: badAnswer
       }
       await this.$store.dispatch('updateTest', test)
+      // this.$router.push(`/tests/${this.$store.state.currentTest._id}`)
       console.log(test)
     }
   }
