+++ package.json
@@ -27,7 +27,6 @@
     "@nuxtjs/dotenv": "^1.4.1",
     "@nuxtjs/pwa": "^3.0.0-0",
     "bcrypt": "^3.0.6",
-    "chart.js": "^2.9.3",
     "cors": "^2.8.4",
     "cross-env": "^5.2.0",
     "esm": "^3.2.25",
@@ -37,10 +36,6 @@
     "morgan": "^1.9.1",
     "nuxt": "^2.0.0",
     "nuxt-fontawesome": "^0.4.0",
-    "vue-chartjs": "^3.5.0",
-    "vue-css-donut-chart": "^1.2.0",
-    "vue-server-renderer": "^2.6.11",
-    "vue-template-compiler": "^2.6.11",
     "vuex-persistedstate": "^2.7.0",
     "w3-css": "^4.1.0"
   },
