+++ pages/tests/index.vue
@@ -1,9 +1,8 @@
 <template>
   <div class>
     <div>
-      <div class="ml-4 mt-5 overline">Los Tests de</div>
-      <div class="ml-4 mb-4 grey--text display-1 ">{{ userName }}</div>
       <div class="header-wraper">
+        <h1 class="ml-4">Los Tests de {{ userName }}</h1>
         <div>
           <v-btn
             class="ma-2"
@@ -31,34 +30,19 @@
         <template v-slot:default>
           <thead>
             <tr>
-              <th class="text-left">Título</th>
+              <th class="text-left">Titulo</th>
               <th class="text-left">Preguntas</th>
               <th class="text-left">Aciertos</th>
               <th class="text-left">Fallos</th>
-              <th class="text-left">En Blanco</th>
               <th class="text-left">% Aciertos</th>
-              <th class="text-left">Estado</th>
             </tr>
           </thead>
           <tbody>
             <tr
-              v-for="item in sorted"
+              v-for="item in tests"
               :key="item._id"
               @click="goToTest(item._id)"
             >
-              <td v-if="!item.desafio">
-                {{ item.title
-                }}<span v-if="item.mostrar_solucion">
-                  <v-icon small right class="red--text">mdi-eye</v-icon></span
-                >
-              </td>
-              <td v-if="!item.desafio">
-                {{ item.no_contestadas.length }}
-              </td>
-              <td v-if="!item.desafio">{{ item.testCheck.right }}</td>
-              <td v-if="!item.desafio">{{ item.testCheck.wrong }}</td>
-              <td v-if="!item.desafio">{{ item.testCheck.blank }}</td>
-              <td v-if="!item.desafio">
               <td
                 v-if="
                   (!item.desafio &&
@@ -102,14 +86,13 @@
                     (!item.desafio && item.time_end === 'ilimitado')
                 "
               >
-
                 {{ (item.testCheck.right / item.no_contestadas.length) * 100 }}
-              </td>
-              <td v-if="!item.desafio">
-                <div v-if="item.time_end === null">
-                  <v-icon>mdi-checkbox-blank-outline</v-icon>
-                </div>
-                <div v-else><v-icon>mdi-check-box-outline</v-icon></div>
+                <span v-if="item.nota === 'suspendido'" class="red--text">{{
+                  item.nota
+                }}</span>
+                <span v-if="item.nota === 'aprobado'" class="green--text">{{
+                  item.nota
+                }}</span>
               </td>
             </tr>
           </tbody>
@@ -193,12 +176,10 @@
 </template>
 
 <script>
-// import API from '~/services/api'
 import { mapGetters } from 'vuex'
 
 export default {
   data() {
-
     return {
       dia_actual: new Date().getDay(),
       mes_actual: new Date().getMonth() + 1,
@@ -206,27 +187,16 @@ export default {
     }
   },
   computed: {
-    ...mapGetters(['tests', 'userName', 'nickName', 'currentTest']),
-    sorted() {
-      let sorted = []
-      for (let i = this.tests.length - 1; i > -1; i--) {
-        let test = this.tests[i]
-        sorted.push(test)
-      }
-      return sorted
-    }
+    ...mapGetters(['tests', 'userName'])
   },
   methods: {
     async testGeneration() {
       await this.$store.dispatch('createTest')
-      this.$router.push(
-        `/tests/${this.$store.state.currentTest._id}/${this.$store.state.currentTest.no_contestadas[0]._id}`
-      )
+      this.$router.push(`/tests/${this.$store.state.currentTest._id}`)
     },
-    async goToTest(id) {
+    goToTest(id) {
       this.$router.push(`/tests/${id}`)
     },
-
     goToTestConfig() {
       this.$router.push(`/tests/config`)
     }
