--- plugins/vue-css-donut-chart.js
@@ -1,5 +1,0 @@
-import Vue from 'vue'
-import Donut from 'vue-css-donut-chart'
-import 'vue-css-donut-chart/dist/vcdonut.css'
-
-Vue.use(Donut)
