--- plugins/vueCharts.js
@@ -1,4 +1,0 @@
-import Vue from 'vue'
-import VueCharts from 'vue-chartjs'
-
-Vue.use(VueCharts)
