+++ services/api.js
@@ -28,6 +28,8 @@ export default {
     })
   },
   generateConfigTest(token, test) {
+    console.log('api')
+    console.log(test)
     return axios
       .post('tests/config', test, { headers: { token } })
       .then(response => {
@@ -35,8 +37,12 @@ export default {
       })
   },
   updateTest(token, data) {
+    let body = {
+      respuestas: data.respuesta,
+      testCheck: data.testCheck
+    }
     return axios
-      .put('tests/' + data.testId, data, { headers: { token } })
+      .put('tests/' + data.testId, body, { headers: { token } })
       .then(response => {
         return response.data
       })
@@ -46,13 +52,6 @@ export default {
   },
   getAllMessageAdmin() {
     return axios.get('messages/').then(response => response.data)
-  },
-  newMessage(token, data) {
-    return axios
-      .post('messages/', data, { headers: { token } })
-      .then(response => {
-        return response.data
-      })
   },
   getTest(testId) {
     return axios.get('tests/' + testId).then(response => response.data)
@@ -93,6 +92,8 @@ export default {
       .then(response => response.data)
   },
   updatePhoto(data) {
+    console.log('api')
+    console.log(data.newPhoto)
     let body = {
       img_url: data.newPhoto
     }
@@ -139,6 +140,7 @@ export default {
     return axios.delete(`temario/${temaId}`).then(response => response.data)
   },
   reply(reply) {
+    console.log(reply)
     const body = {
       respuesta_leida: true,
       respuesta: reply.respuesta
@@ -151,6 +153,7 @@ export default {
     return axios.post(`questions/`, newQuestion).then(response => response.data)
   },
   updateEnunciado(newEnunciado) {
+    // console.log(newEnunciado)
     const body = {
       enunciado: newEnunciado.enunciado
     }
@@ -159,19 +162,12 @@ export default {
       .put('questions/' + newEnunciado.id, body)
       .then(response => response.data)
   },
-  updateQuestionPhoto(photo) {
-    const body = {
-      imagen_url: photo.imagen_url
-    }
-
-    return axios
-      .put('questions/' + photo.id, body)
-      .then(response => response.data)
-  },
   updateOpcion(newOpcion) {
+    // console.log(newOpcion)
     const body = {
       answers: newOpcion.answers
     }
+    console.log(body)
     return axios
       .put('questions/' + newOpcion.id, body)
       .then(response => response.data)
@@ -217,6 +213,7 @@ export default {
       .then(response => response.data)
   },
   change_active(change) {
+    console.log(change)
     const body = {
       active: change.active
     }
@@ -224,6 +221,7 @@ export default {
     return axios.put('users/' + change.id, body).then(response => response.data)
   },
   change_suscription(change) {
+    console.log(change)
     const body = {
       suscription_type: change.suscription_type
     }
@@ -245,7 +243,9 @@ export default {
       return response.data
     })
   },
-  prueba() {},
+  prueba() {
+    console.log('hola desde api.js')
+  },
   updateQuestionById(newQuestion) {
     return axios
       .put('questions/' + newQuestion._id, newQuestion)
