+++ store/index.js
@@ -20,7 +20,6 @@ export const state = () => ({
   tests: [],
   messages: [],
   currentTest: {},
-  currentTestQuestion: {},
   question: {},
   position: '',
   updatePregunta: {},
@@ -60,9 +59,6 @@ export const getters = {
   },
   currentTest(state) {
     return state.currentTest
-  },
-  currentTestQuestion(state) {
-    return state.currentTestQuestion
   },
   question(state) {
     return state.question
@@ -156,9 +152,6 @@ export const mutations = {
   saveCurrentTest(state, test) {
     state.currentTest = test
   },
-  saveCurrentQuestion(state, question) {
-    state.currentTestQuestion = question
-  },
   saveQuestion(state, question) {
     state.question = question
   },
@@ -231,7 +224,9 @@ export const mutations = {
   saveUpdatePregunta(state, pregunta) {
     state.updatePregunta = pregunta
   },
-  prueba() {},
+  prueba() {
+    console.log('console de prueba')
+  },
   evaluar(state) {
     for (let i = 0; i < state.tests.length; i++) {
       if (state.tests[i].nota === 'aprobado') {
@@ -281,11 +276,6 @@ export const actions = {
     }
     return response
   },
-  async newMessage({ commit, state }, message) {
-    await API.newMessage(state.token, message)
-    const mensajes = await API.getAllMessageById(state.userId)
-    commit('saveMessage', mensajes)
-  },
   async createTest({ commit, state }) {
     const newTest = await API.generateTest(state.token)
     commit('saveCurrentTest', newTest)
@@ -293,6 +283,7 @@ export const actions = {
     commit('saveTests', tests)
   },
   async createTestConfig({ commit, state }, testData) {
+    console.log(testData)
     const newTest = await API.generateConfigTest(state.token, testData)
     commit('saveCurrentTest', newTest)
     const tests = await API.getAllTestById(state.userId)
@@ -306,20 +297,20 @@ export const actions = {
       commit('saveTests', tests)
     }
   },
-  // async verRespuesta({ state }, responseBody) {
-  //   const respuesta =
-  //     state.currentTest.no_contestadas[0].answer_wrong[responseBody.number]
-  //   const correcta = state.currentTest.no_contestadas[0].answers_correct
-  //   const enunciado = state.currentTest.no_contestadas[0].enunciado
+  async verRespuesta({ state }, responseBody) {
+    const respuesta =
+      state.currentTest.no_contestadas[0].answer_wrong[responseBody.number]
+    const correcta = state.currentTest.no_contestadas[0].answers_correct
+    const enunciado = state.currentTest.no_contestadas[0].enunciado
 
-  //   if (responseBody.enunciado === enunciado) {
-  //     if (respuesta === correcta) {
-  //       return console.log(true)
-  //     } else {
-  //       return console.log(false)
-  //     }
-  //   }
-  // },
+    if (responseBody.enunciado === enunciado) {
+      if (respuesta === correcta) {
+        return console.log(true)
+      } else {
+        return console.log(false)
+      }
+    }
+  },
   async updateName({ commit, state }, newName) {
     let data = {
       userId: state.userId,
@@ -381,6 +372,7 @@ export const actions = {
     }
   },
   async updateEmail({ commit, state }, newEmail) {
+    console.log(state.userId)
     let data = {
       userId: state.userId,
       newEmail: newEmail
@@ -444,15 +436,6 @@ export const actions = {
   },
   async updateEnunciado({ commit }, body) {
     const response = await API.updateEnunciado(body)
-    if (!response.error) {
-      const pregunta = await API.getQuestionById(body.id)
-      if (!pregunta.error) {
-        commit('saveUpdatePregunta', pregunta)
-      }
-    }
-  },
-  async updateQuestionPhoto({ commit }, body) {
-    const response = await API.updateQuestionPhoto(body)
     if (!response.error) {
       const pregunta = await API.getQuestionById(body.id)
       if (!pregunta.error) {
